Webserver

+++ RTWWebServer/Database/Data/Account.cs
@@ -1,8 +1,8 @@
 namespace RTWWebServer.Database.Data;
 
-public class Account(int id, string userName, string email, string password, string salt)
+public class Account(long id, string userName, string email, string password, string salt)
 {
-    public int Id { get; set; } = id;
+    public long Id { get; set; } = id;
     public string UserName { get; set; } = userName;
     public string Email { get; set; } = email;
     public string Password { get; set; } = password;
